+++ CHANGELOG.md
@@ -16,11 +16,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## 2019-11-22
+
+### Changed [User-Agent string output]
+
+* Changed the output of the user agent string to display platform information with double-dashed separated key-value pairs.
+* Added `UserAgent` parameter to `Connect-Rubrik` that allows specifying additional user-agent information.
+* Added `UserAgentHash` parameter to private function `New-UserAgentString` that allows specifying additional user-agent information.
+* Addresses [Issue 517](https://github.com/rubrikinc/rubrik-sdk-for-powershell/issues/517)
+
 ## 2019-11-20
 
 ### Changed [Made link to quick-start.md relative]
 
-* The link to `quick-start.md` in the readme.md has been updated to a relative link
+* The link to `quick-start.md` in the `readme.md` has been updated to a relative link
 
 ## 2019-11-18